+++ BaragonService/src/main/java/com/hubspot/baragon/service/config/BaragonServiceDWSettings.java
@@ -1,0 +1,19 @@
+package com.hubspot.baragon.service.config;
+
+public class BaragonServiceDWSettings {
+  private final int port;
+  private final String contextPath;
+
+  public BaragonServiceDWSettings(int port, String contextPath) {
+    this.port = port;
+    this.contextPath = contextPath;
+  }
+
+  public int getPort() {
+    return port;
+  }
+
+  public String getContextPath() {
+    return contextPath;
+  }
+}
